requirements.txt
attrs==22.1.0
captcha==0.4
certifi==2022.9.24
charset-normalizer==2.1.1
exceptiongroup==1.0.1
idna==3.4
iniconfig==1.1.1
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
pyparsing==3.0.9
pyTelegramBotAPI==4.7.1
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
tomli==2.0.1
urllib3==1.26.12
